update
# Conflicts:
#	app/Http/Controllers/Inventory/StockController.php
#	app/Http/Controllers/MustReorderStockController.php
#	app/Http/Controllers/StockController.php
#	database/seeders/TransactionSeeder.php
#	routes/web.php

+++ app/Http/Controllers/Inventory/StockWithdrawalController.php
@@ -9,42 +9,59 @@ use App\Models\ItemCatalog;
 use App\Models\Stock;
 use App\Models\StockWithdrawal;
 use App\Models\StockWithdrawalItem;
+use App\Support\HelperService\HandleFileUploadService;
 use App\Support\Inventory\StockWithdrawal\Service\StockWithdrawalService;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Session;
 use Illuminate\View\View;
 use Throwable;
 
 #[AllowDynamicProperties] class StockWithdrawalController extends Controller
 {
 
+
     public function __construct()
     {
         $this->stockWithdrawalService = new StockWithdrawalService();
+        $this->handleUploadService = new HandleFileUploadService();
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function index(): View
     {
+        $this->authorize('view', StockWithdrawal::class);
         return view('pages.inventory.stock-withdrawals.index');
     }
 
 
-    public function data(): JsonResponse
+    /**
+     * @throws AuthorizationException
+     */
+    public function data(Request $request): JsonResponse
     {
-        return response()->json($this->stockWithdrawalService->data());
+        $this->authorize('view', StockWithdrawal::class);
+        return response()->json($this->stockWithdrawalService->data($request));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function show(StockWithdrawal $stockWithdrawal): JsonResponse
     {
-        $stockWithdrawal->load('stockWithdrawalItems', 'stockWithdrawalByEmployees', 'stockWithdrawalItems.stock.item', 'stockWithdrawalByEmployees.user.roles', 'pic.roles', 'stocker.roles');
+        $this->authorize('view', StockWithdrawal::class);
+        $stockWithdrawal->load('stockWithdrawalItems', 'stockWithdrawalByEmployees', 'stockWithdrawalItems.stock.transaction.item', 'stockWithdrawalByEmployees.user.roles', 'pic.roles', 'stocker.roles');
 
         $stockWithdrawalItem = $stockWithdrawal->stockWithdrawalItems->map(function ($item) {
             return [
-                'item_name' => $item->stock->item->name,
+                'item_name' => $item->stock->transaction?->item?->name ?? $item->stock->initialInventoryBalance?->item->name,
                 'code' => $item->code,
                 'qty' => $item->qty,
                 'status' => $item->status,
@@ -70,20 +87,49 @@ use Throwable;
     }
 
 
-    public function search(Request $request)
+    /**
+     * @throws AuthorizationException
+     */
+    public function search(Request $request): JsonResponse
     {
-
+        $this->authorize('view', StockWithdrawal::class);
+        return response()->json($this->stockWithdrawalService->search($request));
     }
 
-    public function filter()
+    /**
+     * @throws AuthorizationException
+     */
+    public function filter(Request $request): JsonResponse
     {
-
+        $this->authorize('view', StockWithdrawal::class);
+        return response()->json($this->stockWithdrawalService->filter($request));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function create(): View
     {
+        $this->authorize('create', StockWithdrawal::class);
         return view('pages.inventory.stock-withdrawals.create');
+    }
+
+
+    /**
+     * @throws AuthorizationException
+     */
+    public function sessionStore(Request $request): void
+    {
+        $this->authorize('create', StockWithdrawal::class);
+        $stockWithdrawal = [
+            'code' => $request->get('code'),
+            'qty' => $request->get('qty'),
+            'stock_id' => $request->get('stock_id'),
+            'item_id' => $request->get('item_id'),
+            'item_name' => $request->get('item_name'),
+        ];
+        session()->push('stock_withdrawal_item', $stockWithdrawal);
     }
 
 
@@ -92,6 +138,7 @@ use Throwable;
      */
     public function store(StockWithdrawalRequest $request): JsonResponse
     {
+        $this->authorize('create', StockWithdrawal::class);
         $this->stockWithdrawalService->store($request);
         return response()->json([
             'message' => 'Data berhasil disimpan'
@@ -105,42 +152,28 @@ use Throwable;
     public function destroy(StockWithdrawal $stockWithdrawal): void
     {
         DB::transaction(function () use ($stockWithdrawal) {
+            $this->authorize('delete', $stockWithdrawal);
             foreach ($stockWithdrawal->stockWithdrawalItems as $item) {
-                if ($item->status === 'Dikembalikan' || $item->status === 'Terpakai') {
+                if ($item->returnedItem) {
                     throw new Exception('Data tidak dapat dihapus, dikarenakan barang sudah ada yg di kembalikan atau terpakai', 403);
                 }
-                ItemCatalog::where('code', $item->code)->update(['status' => 'Tersedia']);
-                $stock = Stock::where('id', $item->stock_id)->first();
-                $stock->increment('qty', $item->qty);
-                $stock->decrement('on_hold_qty', $item->qty);
+                $itemCatalog = ItemCatalog::where('code', $item->code)->first();
+                if ($itemCatalog) {
+                    $itemCatalog->update(['status' => 'Tersedia']);
+                    $itemCatalog->increment('available_qty', $item->qty);
                 }
+                $stock = Stock::where('id', $item->stock_id)->first();
+                $stock->increment('available_qty', $item->qty);
+                $stock->decrement('on_hold_qty', $item->qty);
+            }
             $stockWithdrawal->delete();
         });
     }
 
 
-    public function confirmedByPIC(StockWithdrawal $stockWithdrawal): JsonResponse
-    {
-        $this->stockWithdrawalService->confirmedByPIC($stockWithdrawal);
-
-        return response()->json([
-            'message' => 'data berhasil dikonfirmasi'
-        ]);
-    }
-
-
-    public function confirmedByStocker(StockWithdrawal $stockWithdrawal): JsonResponse
-    {
-        $this->stockWithdrawalService->confirmedByStocker($stockWithdrawal);
-        return response()->json([
-            'message' => 'data berhasil dikonfirmasi'
-        ]);
-    }
-
-
     public function return(StockWithdrawal $stockWithdrawal): View
     {
-        return view('pages.inventory.stock-withdrawals.returned-stock-form', compact('stockWithdrawal'));
+        return view('pages.inventory.stock-withdrawals.returned-item.index', compact('stockWithdrawal'));
     }
 
 
@@ -151,50 +184,30 @@ use Throwable;
 
     public function getStockWithdrawalItem(StockWithdrawalItem $stockWithdrawalItem): JsonResponse
     {
-        return response()->json($stockWithdrawalItem);
+        $itemCatalog = ItemCatalog::with('stock.transaction.item', 'stock.initialInventoryBalance.item')
+            ->where('code', $stockWithdrawalItem->code)
+            ->first();
+        return response()->json([
+            'withdrawal_item' => $stockWithdrawalItem,
+            'item_catalog' => $itemCatalog
+        ]);
+    }
+
+    public function getSessions(): JsonResponse
+    {
+        $stock = session()->get('stock_withdrawal_item') ?? [];
+        return response()->json($stock);
     }
 
 
-    /**
-     * @throws Throwable
-     */
-    public function returningItems(Request $request, StockWithdrawalItem $stockWithdrawalItem): void
+    public function flushSessions(): void
     {
-        $stockWithdrawalItem->load('stock');
-        DB::transaction(function () use ($request, $stockWithdrawalItem) {
-            if ($stockWithdrawalItem->code) {
-                $stockWithdrawalItem->stock->decrement('on_hold_qty', $stockWithdrawalItem->qty);
-                $itemCatalog = ItemCatalog::where('code', $stockWithdrawalItem?->code)->first();
-                ItemCatalog::create([
-                    'transaction_id' => $stockWithdrawalItem->stock?->transaction_id,
-                    'stock_id' => $stockWithdrawalItem->stock?->id,
-                    'draft_stock_id' => $stockWithdrawalItem->stock?->draft_stock_id,
-                    'item_id' => $stockWithdrawalItem->stock?->item_id,
-                    'code' => $itemCatalog->code,
-                    'condition' => $itemCatalog->condition,
-                    'created_by' => $itemCatalog->created_by,
-                    'status' => $request->input('status') === 'Terpakai' ? 'Terpakai' : 'Tersedia',
-                    'initial_balance_inventory_id' => $stockWithdrawalItem->stock?->initial_balance_inventory_id,
-                    'asset_id' => $itemCatalog->asset_id,
-                ]);
-                $itemCatalog->delete();
-                $stockWithdrawalItem->update([
-                    'status' => $request->input('status')
-                ]);
-                if ($request->input('status') === 'Dikembalikan') {
-                    $stockWithdrawalItem->stock->increment('qty');
-                }
-            } else {
-                $stockWithdrawalItem->stock->decrement('on_hold_qty', $stockWithdrawalItem->qty);
-                $stockWithdrawalItem->update([
-                    'qty_used' => $stockWithdrawalItem->qty - $request->input('qty'),
-                    'qty' => $request->input('qty'),
-                ]);
-                $stockWithdrawalItem->stock->increment('qty', $request->input('qty'));
-                $stockWithdrawalItem->update([
-                    'status' => 'Habis',
-                ]);
-            }
-        });
+        session()->forget('stock_withdrawal_item');
+    }
+
+
+    public function deleteSessions(Request $request): void
+    {
+        Session::forget("stock_withdrawal_item.$request->index");
     }
 }
